+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # VRFSearchAndBackup.py #
 ---
 
+## VRFSearchAndBackup v0.0.3-alpha (2013-08-26) ##
+* Created README.md, VRFSearchAndBackup.png
+* Updated TODO.md, code comments to reflect current functionality.  
+* Removed unused modules.
+
 ## VRFSearchAndBackup v0.0.2-alpha (2013-08-22) ##
 * Added whitespace to the end of the "show running-config | section SMVPN"
   command to prevent incorrect matches.  (Without the space at the end, a